refactor: api 모듈 설정, core, infra 구조 변경
