+++ api/build.gradle
@@ -18,9 +18,6 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-security'
     implementation 'org.springframework.boot:spring-boot-starter-oauth2-client'
 
-    // WebFlux (WebClient)
-    implementation 'org.springframework.boot:spring-boot-starter-webflux'
-
     // Mail
     implementation 'org.springframework.boot:spring-boot-starter-mail'
