--- api/src/main/java/com/umust/dobonglife/global/common/webclient/business/dto/response/GeoPointResponse.java
@@ -1,3 +1,0 @@
-package com.umust.dobonglife.global.common.webclient.business.dto.response;
-
-public record GeoPointResponse(double latitude, double longitude) {}
