+++ api/src/main/java/com/umust/dobonglife/global/web/WebConfig.java
@@ -1,26 +1,15 @@
 package com.umust.dobonglife.global.web;
 
-import com.umust.dobonglife.global.auth.resolver.CurrentUserIdArgumentResolver;
-import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.MediaType;
 import org.springframework.http.converter.HttpMessageConverter;
 import org.springframework.http.converter.json.MappingJackson2HttpMessageConverter;
-import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 import java.util.List;
 
 @Configuration
-@RequiredArgsConstructor
 public class WebConfig implements WebMvcConfigurer {
-
-    private final CurrentUserIdArgumentResolver CurrentUserIdArgumentResolver;
-
-    @Override
-    public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
-        resolvers.add(CurrentUserIdArgumentResolver);
-    }
 
     @Override
     public void extendMessageConverters(List<HttpMessageConverter<?>> converters) {
