+++ core/src/main/java/com/umust/dobonglife/global/common/response/BaseErrorResponse.java
@@ -7,24 +7,24 @@ import lombok.Getter;
 import java.time.LocalDateTime;
 
 @Getter
-@JsonPropertyOrder({"success", "code", "message", "timestamp"})
+@JsonPropertyOrder({"success", "status", "message", "timestamp"})
 public class BaseErrorResponse {
 
     private final boolean success;
-    private final int code;
+    private final int status;
     private final String message;
     private final LocalDateTime timestamp;
 
     public BaseErrorResponse(ErrorCode errorCode) {
         this.success = false;
-        this.code = errorCode.getHttpStatus().value();
+        this.status = errorCode.getHttpStatus().value();
         this.message = errorCode.getMessage();
         this.timestamp = LocalDateTime.now();
     }
 
     public BaseErrorResponse(ErrorCode errorCode, String customMessage) {
         this.success = false;
-        this.code = errorCode.getHttpStatus().value();
+        this.status = errorCode.getHttpStatus().value();
         this.message = customMessage;
         this.timestamp = LocalDateTime.now();
     }
