+++ core/src/main/java/com/umust/dobonglife/global/error/CommonErrorCode.java
@@ -1,0 +1,19 @@
+package com.umust.dobonglife.global.error;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import org.springframework.http.HttpStatus;
+
+@Getter
+@AllArgsConstructor
+public enum CommonErrorCode implements ErrorCode {
+
+    BAD_REQUEST(HttpStatus.BAD_REQUEST, "잘못된 요청입니다."),
+    REQUIRED_FIELD_MISSING(HttpStatus.BAD_REQUEST, "필수 입력값이 비어있습니다."),
+    API_NOT_FOUND(HttpStatus.NOT_FOUND, "존재하지 않는 API입니다."),
+    METHOD_NOT_ALLOWED(HttpStatus.METHOD_NOT_ALLOWED, "지원하지 않는 HTTP 메서드입니다."),
+    INTERNAL_SERVER_ERROR(HttpStatus.INTERNAL_SERVER_ERROR, "서버 오류가 발생했습니다.");
+
+    private final HttpStatus httpStatus;
+    private final String message;
+}
