+++ infra/build.gradle
@@ -22,4 +22,7 @@ dependencies {
 
     // Firebase
     implementation 'com.google.firebase:firebase-admin:9.7.1'
+
+    // WebFlux (WebClient)
+    implementation 'org.springframework.boot:spring-boot-starter-webflux'
 }
