+++ infra/src/main/java/com/umust/dobonglife/infra/firebase/FirebaseConfig.java
@@ -3,7 +3,7 @@ package com.umust.dobonglife.infra.firebase;
 import com.google.auth.oauth2.GoogleCredentials;
 import com.google.firebase.FirebaseApp;
 import com.google.firebase.FirebaseOptions;
-import com.umust.dobonglife.global.error.ErrorCode;
+import com.umust.dobonglife.infra.error.InfraErrorCode;
 import com.umust.dobonglife.global.error.exception.BusinessException;
 import jakarta.annotation.PostConstruct;
 import lombok.extern.slf4j.Slf4j;
@@ -52,7 +52,7 @@ public class FirebaseConfig {
             }
         } catch (IOException e) {
             log.error("Firebase 초기화 중 오류 발생: {}", e.getMessage());
-            throw new BusinessException(ErrorCode.FIREBASE_INITIALIZATION_FAILED);
+            throw new BusinessException(InfraErrorCode.FIREBASE_INITIALIZATION_FAILED);
         }
     }
 }
