+++ infra/src/main/java/com/umust/dobonglife/infra/webclient/business/dto/response/GeoPointResponse.java
@@ -1,0 +1,3 @@
+package com.umust.dobonglife.infra.webclient.business.dto.response;
+
+public record GeoPointResponse(double latitude, double longitude) {}
